refactor: Update template imports and restructure design tokens
- Replaced old import paths for design tokens and PDF wrapper components with new paths under the shared directory.
- Introduced new shared design tokens and fonts registration files to enhance modularity and maintainability.
- Updated various components in the resume and cover letter templates to utilize the new shared imports, improving code organization.
