+++ src/templates/modern/resume/components/Experience.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Text, View, StyleSheet, Link } from '@react-pdf/renderer';
-import { colors, spacing } from '@design-tokens';
+import { colors, spacing } from '@template-core/design-tokens';
 import type { ExperienceItem, ResumeSchema } from '@types';
 
 const styles = StyleSheet.create({
