+++ src/templates/modern/resume/components/Skills.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text, View, StyleSheet } from '@react-pdf/renderer';
 
-import { colors, spacing } from '@design-tokens';
+import { colors, spacing } from '@template-core/design-tokens';
 import type { ResumeSchema } from '@types';
 
 const styles = StyleSheet.create({
