+++ src/templates/modern/resume/index.tsx
@@ -8,7 +8,7 @@ import Languages from './components/Languages';
 import Education from './components/Education';
 import Experience from './components/Experience';
 
-import { colors, spacing, typography } from '@design-tokens';
+import { colors, spacing, typography } from '@template-core/design-tokens';
 import type { ResumeSchema, ReactPDFProps } from '@types';
 
 /**
