+++ src/templates/shared/fonts-register.ts
