+++ src/templates/shared/libs/tailwind-colors.ts
