+++ src/templates/shared/with-pdf-wrapper.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Document, Page, View, Text, StyleSheet } from '@react-pdf/renderer';
-import { colors, typography, spacing } from '@design-tokens';
+import { colors, typography, spacing } from './design-tokens';
 
 export type DocumentConfig<T> = {
   getDocumentProps: (data: T) => {
