+++ vite.config.ts
@@ -10,14 +10,8 @@ export default defineConfig({
     alias: {
       '@': fileURLToPath(new URL('./src', import.meta.url)),
       '@ui': fileURLToPath(new URL('./src/ui', import.meta.url)),
-      '@design-tokens': fileURLToPath(new URL('./src/templates/design-tokens.ts', import.meta.url)),
+      '@template-core': fileURLToPath(new URL('./src/templates/shared', import.meta.url)),
       '@types': fileURLToPath(new URL('./src/types.ts', import.meta.url)),
-      '@fonts-register': fileURLToPath(
-        new URL('./src/templates/fonts-register.ts', import.meta.url),
-      ),
-      '@with-pdf-wrapper': fileURLToPath(
-        new URL('./src/templates/with-pdf-wrapper.tsx', import.meta.url),
-      ),
       '@utils': fileURLToPath(new URL('./src/utils', import.meta.url)),
     },
   },
